revenue_management/requirements.txt
# Dependencias principales
polars==0.20.2
streamlit==1.31.0
plotly==5.18.0
matplotlib==3.8.2
openpyxl==3.1.2
prophet==1.1.5
pyyaml==6.0.1

# Dependencias para el desarrollo
pytest==7.4.3
black==23.11.0
flake8==6.1.0